Move slices to services/ folder
