+++ src/components/activities/ActivityCard.jsx
@@ -1,8 +1,8 @@
-import { useUpdateActivityMutation } from "./activitySlice";
+import { useUpdateActivityMutation } from "../../services/activitySlice";
 import { useState, useEffect } from "react";
 
 import { Link } from "react-router-dom";
-import { useGetMeQuery } from "../users/userSlice";
+import { useGetMeQuery } from "../../services/userSlice";
 
 export default function ActivityCard({ activity }) {
   const { data: user, isLoading, error } = useGetMeQuery();
